# api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LabSyslog REST API",
    "version": "v1",
    "description": "Native /v1 adapter over app.Service. Errors are application/problem+json with type https://labsyslog.dev/errors/{code}."
  },
  "paths": {
...
    "/v1/state:validate": {
      "post": {
        "operationId": "state.validate",
        "description": "scope syslog.admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success"
          },
          "default": {
            "description": "application/problem+json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "examples": [
              "https://labsyslog.dev/errors/validation_failed"
            ]
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      }
    }
  }
}